refactor(main): consolidate helpers, extract registrars, drop dead code
- tree:createLib now delegates to allocateAndRegisterLib instead of
  reimplementing it inline; toNamespaceVariable/toNamespaceInspectorNode
  merged into one toNamespaceDescriptor converter.
- config.ts parseConfig rewritten table-driven (~170 lines of per-field
  boilerplate -> four typed key tables). Fixes lastUpdateCheck being
  silently dropped on every load, which reset the auto-updater's check
  throttle on each launch; malformed values are dropped, not fatal.
- module-manager: `installPath ?? path.join(packagesRoot, id)` (11x)
  extracted to moduleDirFor().
- index.ts re-accretion reversed: autosave handlers + snapshot routines
  moved to ipc-register-autosave.ts (returns an AutosaveController for
  the kernel registrar's restart/recovery wiring), environment discovery/
  Packages-tab/installModule handlers moved to ipc-register-environment.ts
  (returns refreshKernelImportCaches for the project-load env sync).
  index.ts: 1427 -> ~950 lines.
- sanitizeScriptName now strips dots and non-identifier characters from
  the stem (tree keys are dot-path segments; a dotted stem corrupted
  path addressing). Mirrors the renderer's CreateTreeItemDialog.
- Dead code deleted: EnvironmentDetector.listAll/installPDV/hasPDVKernel/
  hasJuliaPDVKernel/detectJuliaEnvironments (+ _probeJulia/_juliaCache/
  DetectedJuliaEnvironment/PDVInstallResult), uv-runner uvPythonInstall/
  uvPythonList, comm-router's legacy 3-arg request() shim, kernel-session
  bootstrap's unused `from pdv import PDVTree`.
